+++ src/Controller/CharacterController.php
@@ -44,4 +44,18 @@ class CharacterController extends AbstractController
         return new JsonResponse($character->toArray());
     }
 
+    // src/Controller/CharacterController.php
+    //  INDEX
+    #[
+        Route('/characters/',
+        name: 'app_character_index',
+        methods: ['GET'])
+    ]
+    public function index(): JsonResponse
+    {
+        $this->denyAccessUnlessGranted('characterIndex', null);
+        $characters = $this->characterService->findAll();
+        return new JsonResponse($characters);
+    }
+
 }