+++ src/Service/CharacterServiceInterface.php
@@ -4,4 +4,7 @@ interface CharacterServiceInterface
 {
     // Creates the character
     public function create();
+
+    // Finds all the characters
+    public function findAll();
 }